Refactor fan light device to generic class
