+++ lib/TuyaOAuth2DeviceWithLight.ts
@@ -1,0 +1,146 @@
+import TuyaOAuth2Device from './TuyaOAuth2Device';
+import { ParsedColourData, TuyaStatus } from '../types/TuyaTypes';
+import { TuyaCommand } from '../types/TuyaApiTypes';
+
+/**
+ * Handles all light-related capabilities, except onoff
+ */
+export default class TuyaOAuth2DeviceWithLight extends TuyaOAuth2Device {
+  LIGHT_DIM_CAPABILITY = 'dim';
+
+  LIGHT_DIM_TUYA_CAPABILITY = 'bright_value';
+  LIGHT_TEMP_TUYA_CAPABILITY = 'temp_value';
+  LIGHT_COLOR_TUYA_CAPABILITY = 'colour_data';
+
+  LIGHT_DIM_TUYA_SPECS = 'tuya_brightness';
+  LIGHT_TEMP_TUYA_SPECS = 'tuya_temperature';
+  LIGHT_COLOR_TUYA_SPECS = 'tuya_colour';
+
+  async onOAuth2Init(): Promise<void> {
+    await super.onOAuth2Init();
+
+    // light capabilities
+    const lightCapabilities = [
+      this.LIGHT_DIM_CAPABILITY,
+      'light_hue',
+      'light_saturation',
+      'light_temperature',
+      'light_mode',
+    ].filter(lightCapability => this.hasCapability(lightCapability));
+
+    if (lightCapabilities.length > 0) {
+      this.registerMultipleCapabilityListener(
+        lightCapabilities,
+        capabilityValues => this.onCapabilitiesLight(capabilityValues),
+        150,
+      );
+    }
+  }
+
+  async onTuyaStatus(status: TuyaStatus, changedStatusCodes: string[]): Promise<void> {
+    await super.onTuyaStatus(status, changedStatusCodes);
+
+    // Light
+    const workMode = status['work_mode'] as 'white' | 'colour' | string | undefined;
+    const lightTemp = status[this.LIGHT_TEMP_TUYA_CAPABILITY] as number | undefined;
+    const lightDim = status[this.LIGHT_DIM_TUYA_CAPABILITY] as number | undefined;
+    const lightColor = status[this.LIGHT_COLOR_TUYA_CAPABILITY] as ParsedColourData | undefined;
+
+    if (workMode === 'white') {
+      await this.safeSetCapabilityValue('light_mode', 'temperature');
+    } else if (workMode === 'colour') {
+      await this.safeSetCapabilityValue('light_mode', 'color');
+    } else {
+      await this.safeSetCapabilityValue('light_mode', null);
+    }
+
+    if (lightTemp) {
+      const specs = this.store[this.LIGHT_TEMP_TUYA_SPECS];
+      const light_temperature = (lightTemp - specs.min) / (specs.max - specs.min);
+      await this.safeSetCapabilityValue('light_temperature', light_temperature);
+    }
+
+    if (lightDim && (workMode === 'white' || workMode === undefined)) {
+      const specs = this.store[this.LIGHT_DIM_TUYA_SPECS];
+      const dim = (lightDim - specs.min) / (specs.max - specs.min);
+      await this.safeSetCapabilityValue(this.LIGHT_DIM_CAPABILITY, dim);
+    }
+
+    if (lightColor) {
+      const specs = this.store[this.LIGHT_COLOR_TUYA_SPECS];
+      const h = (lightColor.h - specs.h.min) / (specs.h.max - specs.h.min);
+      const s = (lightColor.s - specs.s.min) / (specs.s.max - specs.s.min);
+
+      await this.safeSetCapabilityValue('light_hue', h);
+      await this.safeSetCapabilityValue('light_saturation', s);
+
+      if (workMode === 'colour') {
+        const v = (lightColor.v - specs.v.min) / (specs.v.max - specs.v.min);
+        await this.safeSetCapabilityValue(this.LIGHT_DIM_CAPABILITY, v);
+      }
+    }
+  }
+
+  async onCapabilitiesLight({
+    light_dim = this.getCapabilityValue(this.LIGHT_DIM_CAPABILITY),
+    light_mode = this.getCapabilityValue('light_mode'),
+    light_hue = this.getCapabilityValue('light_hue'),
+    light_saturation = this.getCapabilityValue('light_saturation'),
+    light_temperature = this.getCapabilityValue('light_temperature'),
+  }): Promise<void> {
+    const commands: TuyaCommand[] = [];
+
+    if (!light_mode) {
+      if (this.hasCapability('light_hue')) {
+        light_mode = 'color';
+      } else {
+        light_mode = 'temperature';
+      }
+    }
+
+    if (this.hasTuyaCapability('work_mode')) {
+      commands.push({
+        code: 'work_mode',
+        value: light_mode === 'color' ? 'colour' : 'white',
+      });
+    }
+
+    if (light_mode === 'color') {
+      const specs = this.store[this.LIGHT_COLOR_TUYA_SPECS];
+      const h = specs.h.min + light_hue * (specs.h.max - specs.h.min);
+      const s = specs.s.min + light_saturation * (specs.s.max - specs.s.min);
+      const v = specs.v.min + light_dim * (specs.v.max - specs.v.min);
+
+      commands.push({
+        code: this.LIGHT_COLOR_TUYA_CAPABILITY,
+        value: { h, s, v },
+      });
+    } else {
+      // Dim
+      if (light_dim && this.hasTuyaCapability(this.LIGHT_DIM_TUYA_CAPABILITY)) {
+        const specs = this.store[this.LIGHT_DIM_TUYA_SPECS];
+        const brightValue = specs.min + light_dim * (specs.max - specs.min);
+
+        commands.push({
+          code: this.LIGHT_DIM_TUYA_CAPABILITY,
+          value: brightValue,
+        });
+      }
+
+      // Temperature
+      if (light_temperature && this.hasTuyaCapability(this.LIGHT_TEMP_TUYA_CAPABILITY)) {
+        const specs = this.store[this.LIGHT_TEMP_TUYA_SPECS];
+        const tempValue = specs.min + light_temperature * (specs.max - specs.min);
+
+        commands.push({
+          code: this.LIGHT_TEMP_TUYA_CAPABILITY,
+          value: tempValue,
+        });
+      }
+    }
+
+    if (commands.length) {
+      await this.sendCommands(commands);
+    }
+  }
+}
